split breeds layout into components - add searchbar functionality with debounce
